+++ Term3/Models/ProjectModel.cs
@@ -13,7 +13,7 @@ namespace TErm.Models
         public string name { get; set; }
         public LinksModel _links { get; set; }
         public List<IssuesModel> issuesList { get; set; }
-        public double projectTime { get; set; } //в секундах
+        public int projectTime { get; set; } //в секундах
 
         public List<SelectListItem> projectsList { get; set; }
     }
